Add metadata to output for replicable analysis
This commit enhances the output file by including
metadata necessary for reproducing the analysis.
It records the first and last commit of each repository,
timestamps for when the analysis started and finished,
current version of grimoirelab-metrics cli, and all
parameters passed to the script along with their
default values if not provided.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>