[metric] Add Contributor Growth
This metric measures the growth of active contributors, defined as the
number of people sending one or more code contributions (in this case,
Git commits) in a given period.

To calculate it, the period is split into two halves, and the number of active
contributors in each half is compared. Growth is the difference between the
second half and the first half.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>